Refactored GQL Guard with CurrentUser Decorator to Extract JWT User Payload
