+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { GraphQLModule } from '@nestjs/graphql';
 import { PubSub } from 'graphql-subscriptions';
 import { UserModule } from './user/user.module';
 import { PrismaService } from './prisma/prisma.service';
-import { GqlAuthGuard } from './app.service';
+import { GqlAuthGuard } from './gql.guard';
 import { AuthModule } from './auth/auth.module';
 import { AuthorModule } from './author/author.module';
 import { BookModule } from './book/book.module';
