--- src/app.service.ts
@@ -1,24 +1,0 @@
-import { Injectable, ExecutionContext } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
-import { GqlExecutionContext } from '@nestjs/graphql';
-import { ExecutionContextHost } from '@nestjs/core/helpers/execution-context-host';
-import { AuthenticationError } from 'apollo-server-core';
-
-@Injectable()
-export class GqlAuthGuard extends AuthGuard('jwt') {
-  canActivate(context: ExecutionContext) {
-    const ctx = GqlExecutionContext.create(context);
-    const { req } = ctx.getContext();
-    // console.log(req);
-
-    return super.canActivate(new ExecutionContextHost([req]));
-  }
-
-  handleRequest(err: any, user: any) {
-    if (err || !user) {
-      throw err ||
-        new AuthenticationError('Auhtorization invalid or not provided!');
-    }
-    return user;
-  }
-}
