+++ src/user/user.decorator.ts
@@ -1,0 +1,9 @@
+import { createParamDecorator, ExecutionContext } from '@nestjs/common';
+import { GqlExecutionContext } from '@nestjs/graphql';
+
+export const CurrentUser = createParamDecorator(
+  (data: string, context: ExecutionContext) => {
+    const ctx = GqlExecutionContext.create(context);
+    return data ? ctx.getContext().req.user && ctx.getContext().req.user[data] : ctx.getContext().req.user;
+  },
+);
